+++ src/rhsmlib/dbus/services/__init__.py
@@ -1,0 +1,2 @@
+from rhsmlib.dbus.services.private_service import PrivateService
+from rhsmlib.dbus.services.base_service import BaseService
