+++ src/rhsmlib/dbus/services/config/__init__.py
@@ -1,0 +1,1 @@
+from rhsmlib.dbus.services.config.config_service import ConfigService
