+++ src/rhsmlib/dbus/services/main/__init__.py
@@ -1,0 +1,1 @@
+from rhsmlib.dbus.services.main.main_service import MainService
