+++ src/rhsmlib/dbus/services/register/__init__.py
@@ -1,0 +1,1 @@
+from rhsmlib.dbus.services.register.register_service import RegisterService
